# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Distributed Lock Lease Manager API
  version: 0.1.0
  description: |
    LS-SPEC-001 準拠のREST API契約。
servers:
  - url: https://api.example.com
...
paths:
  /leases/acquire:
    post:
      operationId: acquireLease
      tags: [leases]
      summary: Acquire a lease
      parameters:
        - $ref: "#/components/parameters/OwnerHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcquireLeaseRequest"
      responses:
        "201":
          description: Lease acquired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaseResponse"
        "409":
          description: Lock held
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: owner_id does not match authenticated identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    OwnerHeader:
      name: x-owner-id
      in: header
      required: false
      schema:
        type: string
      description: |
        認証トークン由来の owner を運ぶヘッダ。指定時は request body の owner_id と一致が必要。
...
  schemas:
    AcquireLeaseRequest:
      $ref: "../schema/lease-acquire-request.schema.json"
...
    LeaseResponse:
      $ref: "../schema/lease-response.schema.json"
...
    ErrorResponse:
      $ref: "../schema/error-response.schema.json"